Add files via upload
The task 4 function is the key aspect of this algorithm.

It takes words and a dictionary as input and spell checks the words to their first, second or third edit distance away.

e.g. if the word exists in the dictionary, then it is output. But if it is not in the dictionary and there is a word that exists 1, 2 or 3 edit distance away, it will output the word with the smallest edit distance.